* fix, pmtelemetryd: stats counters moved to u_int64_t

+++ src/bgp/bgp.h
@@ -143,10 +143,10 @@ struct bgp_xconnects {
 };
 
 struct bgp_peer_stats {
-    u_int32_t packets; /* Datagrams received */
-    u_int32_t packet_bytes; /* Bytes read off the socket */
-    u_int32_t msg_bytes; /* Bytes in the decoded messages */
-    u_int32_t msg_errors; /* Errors detected in message content */
+    u_int64_t packets; /* Datagrams received */
+    u_int64_t packet_bytes; /* Bytes read off the socket */
+    u_int64_t msg_bytes; /* Bytes in the decoded messages */
+    u_int64_t msg_errors; /* Errors detected in message content */
     time_t last_check; /* Timestamp when stats were last checked */
 };
